refactor: pure values
