+++ packages/mask/src/components/InjectedComponents/ProfileTabContent.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from 'react'
 import { useUpdateEffect } from 'react-use'
 import { first } from 'lodash-unified'
-import type { NextIDPlatform } from '@masknet/shared-base'
+import { NextIDPlatform, EMPTY_LIST } from '@masknet/shared-base'
 import { Box, CircularProgress } from '@mui/material'
 import { makeStyles, useStylesExtends } from '@masknet/theme'
 import { useAddressNames } from '@masknet/web3-shared-evm'
@@ -13,7 +13,6 @@ import { useCurrentVisitingIdentity, useLastRecognizedIdentity } from '../DataSo
 import { useNextIDBoundByPlatform } from '../DataSource/useNextID'
 import { usePersonaConnectStatus } from '../DataSource/usePersonaConnectStatus'
 import { activatedSocialNetworkUI } from '../../social-network'
-import { EMPTY_LIST } from '../../../utils-pure'
 
 function getTabContent(tabId: string) {
     return createInjectHooksRenderer(useActivatedPluginsSNSAdaptor.visibility.useAnyMode, (x) => {
