+++ packages/mask/src/plugins/NextID/components/Tip/TipButton.tsx
@@ -1,9 +1,8 @@
 import { TipCoin } from '@masknet/icons'
 import { usePostInfoDetails } from '@masknet/plugin-infra'
-import { NextIDPlatform, ProfileIdentifier } from '@masknet/shared-base'
+import { EMPTY_LIST, NextIDPlatform, ProfileIdentifier } from '@masknet/shared-base'
 import { makeStyles, ShadowRootTooltip } from '@masknet/theme'
 import { queryExistedBindingByPersona, queryIsBound } from '@masknet/web3-providers'
-import { EMPTY_LIST } from '@masknet/web3-shared-evm'
 import classnames from 'classnames'
 import { uniq } from 'lodash-unified'
 import { FC, HTMLProps, MouseEventHandler, useCallback, useMemo } from 'react'
