+++ packages/mask/src/plugins/NextID/components/Tip/TipDialog.tsx
@@ -1,8 +1,9 @@
 import { SuccessIcon } from '@masknet/icons'
 import { PluginId, useActivatedPlugin, usePluginIDContext } from '@masknet/plugin-infra'
 import { NFTCardStyledAssetPlayer } from '@masknet/shared'
+import { EMPTY_LIST } from '@masknet/shared-base'
 import { makeStyles } from '@masknet/theme'
-import { EMPTY_LIST, TransactionStateType, useChainId, useERC721TokenDetailed } from '@masknet/web3-shared-evm'
+import { TransactionStateType, useChainId, useERC721TokenDetailed } from '@masknet/web3-shared-evm'
 import { DialogContent, Typography } from '@mui/material'
 import { useCallback, useEffect, useMemo } from 'react'
 import { useBoolean } from 'react-use'
