+++ packages/mask/src/plugins/Savings/SNSAdaptor/SavingsDialog.tsx
@@ -2,9 +2,8 @@ import { useState, useMemo } from 'react'
 import { useAsync } from 'react-use'
 import { Typography, DialogContent } from '@mui/material'
 import { ChainId, getChainIdFromNetworkType, useChainId } from '@masknet/web3-shared-evm'
-import { isDashboardPage } from '@masknet/shared-base'
+import { isDashboardPage, EMPTY_LIST } from '@masknet/shared-base'
 import { useI18N } from '../../../utils'
-import { EMPTY_LIST } from '../../../../utils-pure'
 import { InjectedDialog } from '../../../components/shared/InjectedDialog'
 import { WalletStatusBox } from '../../../components/shared/WalletStatusBox'
 import { AllProviderTradeContext } from '../../Trader/trader/useAllProviderTradeContext'
