+++ packages/mask/src/plugins/Trader/trader/uniswap/usePairs.ts
@@ -1,16 +1,16 @@
 import { useMemo } from 'react'
 import { useAsyncRetry } from 'react-use'
+import { numberToHex } from 'web3-utils'
 import { Pair } from '@uniswap/v2-sdk'
 import { CurrencyAmount, Token } from '@uniswap/sdk-core'
+import { EMPTY_LIST } from '@masknet/shared-base'
 import { useMultipleContractSingleData } from '@masknet/web3-shared-evm'
 import { getPairAddress } from '../../helpers'
 import { usePairContracts } from '../../contracts/uniswap/usePairContract'
 import type { TradeProvider } from '@masknet/public-api'
 import { useGetTradeContext } from '../useGetTradeContext'
 import { TargetChainIdContext } from '../useTargetChainIdContext'
-import { numberToHex } from 'web3-utils'
 import { useTargetBlockNumber } from '../useTargetBlockNumber'
-import { EMPTY_LIST } from '../../../../../utils-pure'
 
 export enum PairState {
     NOT_EXISTS = 0,
