+++ packages/shared-base/src/index.ts
@@ -7,6 +7,7 @@ export * from './i18n'
 export * from './utils'
 export * from './kv-storage'
 export * from './crypto'
+export * from './Pure'
 export * from './Persona/type'
 export * from './Post/type'
 export * from './Persona/type'
