+++ packages/shared-base/tsconfig.json
@@ -7,5 +7,5 @@
   },
   "include": ["./src", "./src/**/*.json"],
   // Only type-level dependency.
-  "references": [{ "path": "../public-api" }, { "path": "../web3-shared/evm" }, { "path": "../typed-message/base" }]
+  "references": [{ "path": "../public-api" }, { "path": "../typed-message/base" }]
 }
