+++ packages/theme/package.json
@@ -22,6 +22,7 @@
   },
   "dependencies": {
     "@masknet/icons": "workspace:*",
+    "@masknet/shared-base": "workspace:*",
     "@masknet/storybook-shared": "workspace:*",
     "@types/qrcode": "^1.4.2",
     "@types/tinycolor2": "^1.4.3",
