+++ packages/theme/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "./dist/src.tsbuildinfo"
   },
   "include": ["./src"],
-  "references": [{ "path": "../icons" }]
+  "references": [{ "path": "../icons" }, { "path": "../shared-base" }]
 }
