+++ packages/web3-shared/base/src/utils/number.ts
@@ -44,6 +44,7 @@ export function multipliedBy(a: BigNumber.Value, b: BigNumber.Value) {
 }
 
 /** 10 ** n */
+/** @deprecated use scale10 */
 export function pow10(n: BigNumber.Value) {
     return new BigNumber(10).pow(n)
 }
