+++ packages/web3-shared/evm/types/index.ts
@@ -1,5 +1,11 @@
 import type { TransactionConfig as TransactionConfig_ } from 'web3-core'
 import type { NonPayableTransactionObject, PayableTransactionObject } from '@masknet/web3-contracts/types/types'
+import type { ProviderType } from '../../../public-api'
+
+// TODO: remove the depends of MaskSettingsEvents
+export { NetworkType, ProviderType, FungibleAssetProvider, NonFungibleAssetProvider } from '../../../public-api'
+
+export type { PriceRecord, CryptoPrice } from '../../../public-api'
 
 export interface SendOverrides {
     chainId?: ChainId
@@ -15,14 +21,6 @@ export enum CurrencyType {
     USD = 'usd',
 }
 
-export interface PriceRecord {
-    [currency: string]: number
-}
-
-/** Base on response of coingecko's token price API */
-export interface CryptoPrice {
-    [token: string]: PriceRecord
-}
 export type ChainIdOptionalRecord<T> = { [k in ChainId]?: T }
 export type ChainIdRecord<T> = { [k in ChainId]: T }
 
@@ -83,40 +81,10 @@ export enum ChainId {
     Conflux = 1030,
 }
 
-export enum ProviderType {
-    MaskWallet = 'Maskbook',
-    MetaMask = 'MetaMask',
-    WalletConnect = 'WalletConnect',
-    Fortmatic = 'Fortmatic',
-    Coin98 = 'Coin98',
-    MathWallet = 'MathWallet',
-    WalletLink = 'WalletLink',
-    CustomNetwork = 'CustomNetwork',
-}
-
 export enum LockStatus {
     INIT = 0,
     UNLOCK = 1,
     LOCKED = 2,
-}
-
-// If you change this enum, please sync it to packages/public-api/src/web.ts
-// (it's a breaking change. Please notify the iOS and Android dev)
-export enum NetworkType {
-    Ethereum = 'Ethereum',
-    Binance = 'Binance',
-    Polygon = 'Polygon',
-    Arbitrum = 'Arbitrum',
-    xDai = 'xDai',
-    Celo = 'Celo',
-    Fantom = 'Fantom',
-    Aurora = 'Aurora',
-    Avalanche = 'Avalanche',
-    Boba = 'Boba',
-    Fuse = 'Fuse',
-    Metis = 'Metis',
-    Optimistic = 'Optimistic',
-    Conflux = 'Conflux',
 }
 
 export interface Wallet {
@@ -577,18 +545,6 @@ export interface Asset {
 export enum DomainProvider {
     ENS = 'ENS',
     UNS = 'UNS',
-}
-
-export enum FungibleAssetProvider {
-    ZERION = 'Zerion',
-    DEBANK = 'Debank',
-}
-
-export enum NonFungibleAssetProvider {
-    OPENSEA = 'OpenSea',
-    RARIBLE = 'Rarible',
-    NFTSCAN = 'NFTScan',
-    ZORA = 'Zora',
 }
 
 export type UnboxTransactionObject<T> = T extends NonPayableTransactionObject<infer R>
