+++ packages/web3-shared/evm/utils/index.ts
@@ -7,5 +7,4 @@ export * from './chainDetailed'
 export * from './transaction'
 export * from './domain'
 export * from './payload'
-export * from './misc'
 export * from './provider'
